Add new exercise

+++ C#OOP/08.MockingAndTestDrivenDevelopment/INStock - Skeleton/INStock/Contracts/IProductStock.cs
@@ -22,7 +22,7 @@
 
         IEnumerable<IProduct> FindAllInRange(decimal lo, decimal hi);
 
-        IEnumerable<IProduct> FindAllByPrice(double price);
+        IEnumerable<IProduct> FindAllByPrice(decimal price);
 
         IEnumerable<IProduct> FindAllByQuantity(int quantity);
     }
